Newzz - A Smart News App
This app gives latest technical news that is fetched from The Next Web source.
Moreover it gives more that 60+ channels/sources to choose from whether it's sports to drama to music. This app covers all.
The user can also sort the news in order of latest, top or trending provided by the source.